+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 
 ## Improvements
-
+- #1572 Setting up default value for max_bytes_chunk_read to 256 MB
 
 ## Bug Fixes
 - #1570 Fix build due to changes in rmm device buffer