+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 
 
 ## Improvements
+- #1571 Update ucx-py versions to 0.21
 - #1572 Setting up default value for max_bytes_chunk_read to 256 MB
+
 
 ## Bug Fixes
 - #1570 Fix build due to changes in rmm device buffer