+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 ## Improvements
 - #1571 Update ucx-py versions to 0.21
+- #1554 return ok for filesystems
 - #1572 Setting up default value for max_bytes_chunk_read to 256 MB