+++ CHANGELOG.md
@@ -10,7 +10,8 @@
 
 
 ## Bug Fixes
-- #1570 Fix build due to changes in rmm device buffer 
+- #1570 Fix build due to changes in rmm device buffer
+- #1576 Fix `CC`/`CXX` variables in CI
 
 
 # BlazingSQL 21.06.00 (June 10th, 2021)