+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - #1570 Fix build due to changes in rmm device buffer
 - #1576 Fix `CC`/`CXX` variables in CI
 - #1581 Fix latest cudf dependencies
+- #1585 Fix for GCS credentials from filepath
 
 
 # BlazingSQL 21.06.00 (June 10th, 2021)