+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - #1576 Fix `CC`/`CXX` variables in CI
 - #1581 Fix latest cudf dependencies
 - #1583 Fix concat suite E2E test for nested calls
+- #1585 Fix for GCS credentials from filepath
 
 
 # BlazingSQL 21.06.00 (June 10th, 2021)